requirements.txt
pillow
psutil
pywin32
keyboard
PyInstaller
pytesseract
PyDirectInput
opencv-python
